run tests sequentially
